requirements.txt
fastapi
uvicorn
weaviate-client
requests